% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllocateBiomass.r
\name{AllocateBiomass}
\alias{AllocateBiomass}
\title{AllocateBiomass}
\usage{
AllocateBiomass(state, site, parms, weather)
}
\arguments{
\item{state}{current model state}

\item{site}{site data}

\item{parms}{model parameters}

\item{weather}{weather data}
}
\value{
updated model state
}
\description{
AllocateBiomass
}
