% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateModifiers.r
\name{CalculateModifiers}
\alias{CalculateModifiers}
\title{CalculateModifiers}
\usage{
CalculateModifiers(state, weather, site, parms, general.info)
}
\arguments{
\item{state}{current model state}

\item{weather}{weather data}

\item{site}{site data}

\item{parms}{model parameters}
}
\value{
updated model state
}
\description{
CalculateModifiers
}
