% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFuncsProjections.R
\name{chess_xy_latlon}
\alias{chess_xy_latlon}
\title{chess_xy_latlon}
\usage{
chess_xy_latlon(x, y)
}
\arguments{
\item{x}{CHESS x coord}

\item{y}{CHESS y coord}
}
\value{
dataframe of lat lon
}
\description{
convert xy vals from CHESS to lat lon
}
