% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewPlantation.r
\name{CreateNewPlantation}
\alias{CreateNewPlantation}
\title{CreateNewPlantation initiate plantation}
\usage{
CreateNewPlantation(state, parms, fma)
}
\arguments{
\item{state}{model state}

\item{parms}{model parameters}
}
\value{
updated models state
}
\description{
CreateNewPlantation initiate plantation
}
