% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNewSprouts.r
\name{CreateNewSprouts}
\alias{CreateNewSprouts}
\title{CreateNewSprouts initiate new sprouts}
\usage{
CreateNewSprouts(state, parms, newsprouts)
}
\arguments{
\item{state}{model state}

\item{parms}{model parameters}
}
\value{
updated models state
}
\description{
CreateNewSprouts initiate new sprouts
}
