% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPriors.R
\name{createPriors_sitka}
\alias{createPriors_sitka}
\title{createPriors create priors for MCMC calibration for sitka}
\usage{
createPriors_sitka(sitka, sd = F)
}
\arguments{
\item{sitka}{current sitka parameters to base priors on}
}
\value{
dataframe of priors
}
\description{
createPriors create priors for MCMC calibration for sitka
}
