% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoThinning.r
\name{DoThinning}
\alias{DoThinning}
\title{DoThinning thin stand}
\usage{
DoThinning(state, parms, fma, presc)
}
\arguments{
\item{state}{model state}

\item{parms}{model parameters}
}
\value{
updated models state
}
\description{
DoThinning thin stand
}
