% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologyFuncs.R
\name{drainageFunc}
\alias{drainageFunc}
\title{drainageFunc calculate volume of water draining out of soil}
\usage{
drainageFunc(parms, weather, SWC, soilVol, K_drain)
}
\arguments{
\item{parms}{global 3PG parameter values}

\item{weather}{global weather values}

\item{SWC}{soil water content of zone to drain}
}
\value{
drainage
}
\description{
Calculates drainage out of a soil zone
}
