% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateAPAR.r
\name{EstimateAPAR}
\alias{EstimateAPAR}
\title{EstimateAPAR}
\usage{
EstimateAPAR(state, weather, parms, general.info)
}
\arguments{
\item{state}{current model state}

\item{weather}{weather data}

\item{parms}{model parameters}

\item{general.info}{general metadata}

\item{site}{site data}
}
\value{
updated model state
}
\description{
EstimateAPAR
}
