% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateMortality.r
\name{EstimateMortality}
\alias{EstimateMortality}
\title{EstimateMortality}
\usage{
EstimateMortality(state, parms)
}
\arguments{
\item{state}{current model state}

\item{parms}{model parameters}
}
\value{
updated model state
}
\description{
EstimateMortality
}
