% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNPP.r
\name{EstimateNPP}
\alias{EstimateNPP}
\title{EstimateNPP}
\usage{
EstimateNPP(state, parms)
}
\arguments{
\item{state}{current model state}

\item{parms}{model parameters}
}
\value{
updated model state
}
\description{
EstimateNPP
}
