% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFuncsProjections.R
\name{ext_eu_soil}
\alias{ext_eu_soil}
\title{ext_eu_soil}
\usage{
ext_eu_soil(data_location, spat_chunk, val_name, max_distance, rast = T)
}
\arguments{
\item{spat_chunk}{spatial chunk file containing grid square locations, climate and metadata}

\item{val_name}{name to use for extracted data}

\item{max_distance}{if closest grid square in soil data to climate data is empty (e.g. water, no data) 
how far to look for next grid square before defaulting to NA}

\item{rast}{is the input file one of the eu raster files or astleys bgs data}

\item{map_location}{file location for eu map (wp,sp,cond etc.)}
}
\description{
function to extract soil data from EU maps and match with closest grid ref of climate
}
