% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateDatFuncs.R
\name{getClimDatX}
\alias{getClimDatX}
\title{getClimDatX get climate data for Harwood - sitka spruce calibrations}
\usage{
getClimDatX(timeStep = "monthly", climDir)
}
\arguments{
\item{timeStep}{time step to get data for ("weekly", "monthly" or "daily")}

\item{climDir}{directory where climate data is stored}
}
\value{
climate data
}
\description{
getClimDatX get climate data for Harwood - sitka spruce calibrations
}
