% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateDatFuncs.R
\name{getClmPine}
\alias{getClmPine}
\title{getClmPine get climate data for finnish scots pine using PREDICT database}
\usage{
getClmPine(timeStep = "monthly")
}
\arguments{
\item{timeStep}{time step to get data for ("weekly", "monthly" or "daily")}
}
\value{
climate data
}
\description{
getClmPine get climate data for finnish scots pine using PREDICT database
}
