% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getParms}
\alias{getParms}
\title{getParms get baseline parameters for sitka spruce}
\usage{
getParms(
  weather = clm_df_full,
  Wl = 0.01,
  WlDormant = 0,
  Wr = 0.01,
  Wsbr = 0.1,
  Wlitt = 0,
  YrC = 47.97,
  YlC = 95.94,
  OC = 335.78,
  YrN = 2.02,
  YlN = 4.05,
  ON = 14.7,
  Nav = 12.4,
  N = 2000,
  rotation = 1,
  cycle = 1,
  rm.sprouts = F,
  nyears = 35,
  initial.month = 1,
  latitude = 57.06,
  soilclass = -1,
  ASW = 165,
  MaxASW = 500,
  MinASW = 0,
  CO2 = 400,
  pFS2 = 0.979729182866315,
  pFS20 = 0.450083704735021,
  aS = 0.041517998716385,
  nS = 2.64745625287154,
  pRx = 0.637320766593578,
  pRn = 0.285132846643722,
  Tmin = -5,
  Topt = 15,
  Tmax = 35,
  kF = 0,
  SWconst0 = 0.3,
  SWpower0 = 4,
  m0 = 0.392548265174109,
  MaxAge = 265.6,
  nAge = 3.545,
  rAge = 0.796,
  gammaFx = 0.01888,
  gammaF0 = 0.001,
  tgammaF = 60,
  Rttover = 0.1,
  MaxCond = 0.02,
  LAIgcx = 1.97882119903727,
  BLcond = 0.102217047783467,
  wSx1000 = 500,
  thinPower = 1.5,
  mF = 0.633672385962458,
  mR = 0.650445331850193,
  mS = 0.2,
  SLA0 = 3.85479046435674,
  SLA1 = 5.33472018885286,
  tSLA = 3.94078881883446,
  k = 0.5,
  fullCanAge = 18,
  MaxIntcptn = 0.1,
  LAImaxIntcptn = 5,
  alpha = 0.06,
  Y = 0.633301670948344,
  poolFractn = 0,
  e20 = 2.2,
  rhoAir = 1.2,
  lambda = 2460000,
  VPDconv = 0.000622,
  fracBB0 = 0.3,
  fracBB1 = 0.1,
  tBB = 10,
  rhoMin = 0.39,
  rhoMax = 0.39,
  tRho = 5,
  Qa = -50,
  Qb = 0.8,
  gDM_mol = 24,
  molPAR_MJ = 2.3,
  CoeffCond = 0.09986604057721,
  fCalpha700 = 1.433,
  fCg700 = 0.451,
  fCalphax = 2.33333333333333,
  fCg0 = 1.75,
  MinCond = 0.005,
  klmax = 0.06903041028657,
  krmax = 0.007021449543068,
  komax = 0.00840935020844,
  hc = 0.494539540825284,
  qir = 148.092408851735,
  qil = 14.9377246340992,
  qh = 40.4572399963215,
  qbc = 7.00578471546651,
  el = 0.739904866285297,
  er = 0.769592209202898,
  Nf = 0.0196791366844,
  Navm = 0.032129816636208,
  Navx = 8.75614468799907,
  leaf.grow = 0,
  leaf.fall = 0,
  Wl.s = 0.01,
  Wsbr.s = 0.1,
  Wr.s = 0.01,
  pWl.sprouts = 0.5,
  pWsbr.sprouts = 0.9,
  cod.pred = "3PG",
  cod.clim = "Month",
  waterBalanceSubMods = T,
  wiltPoint = 0.14,
  fieldCap = 0.2,
  satPoint = 0.35,
  K_s = 0.5,
  shared_area = 4,
  V_nr = 0.5,
  maxRootDepth = 0.5,
  sigma_zR = 0.3,
  SWC_nr = 10,
  E_S1 = 0.3,
  E_S2 = 0.6,
  MaxASW_state = 50,
  K_drain = 0.5,
  timeStp = 12,
  rainMod = 1,
  tempMod = 1,
  fsMod = 1,
  startN = 6,
  startC = 200,
  pseudo = F,
  Q10 = 4,
  Q10X = 0,
  Nleach_r = 0,
  K_drain_nrz = 0.8
)
}
\value{
list of parameters and values
}
\description{
getParms get baseline parameters for sitka spruce
}
