% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParmsOak.R
\name{getParmsOak}
\alias{getParmsOak}
\title{getParmsOak get baseline parameters for oak (Species interactions under climate change in mixed stands of Scots pine and pedunculate oak Bouwman et al. 2021)}
\usage{
getParmsOak(
  weather = clm_df_full,
  Wl = 0.01,
  WlDormant = 0,
  Wr = 0.01,
  Wsbr = 0.1,
  Wlitt = 0,
  YrC = 47.97,
  YlC = 95.94,
  OC = 335.78,
  YrN = 2.02,
  YlN = 4.05,
  ON = 14.7,
  Nav = 12.4,
  N = 2000,
  rotation = 1,
  cycle = 1,
  rm.sprouts = F,
  nyears = 35,
  initial.month = 1,
  latitude = 57.06,
  soilclass = -1,
  ASW = 165,
  MaxASW = 500,
  MinASW = 0,
  CO2 = 400,
  pFS2 = 0.7,
  pFS20 = 0.08,
  aS = 0.06,
  nS = 2.74,
  pRx = 0.7,
  pRn = 0.2,
  Tmin = -5,
  Topt = 20,
  Tmax = 25,
  kF = 0,
  SWconst0 = 0.3,
  SWpower0 = 4,
  m0 = 0,
  MaxAge = 350,
  nAge = 4,
  rAge = 0.95,
  gammaFx = 0.02,
  gammaF0 = 0.001,
  tgammaF = 60,
  Rttover = 0.015,
  MaxCond = 0.02,
  LAIgcx = 3.33,
  BLcond = 0.2,
  wSx1000 = 500,
  thinPower = 1.5,
  mF = 0.633672385962458,
  mR = 0.650445331850193,
  mS = 0.2,
  SLA0 = 18.49,
  SLA1 = 14.62,
  tSLA = 7.35,
  k = 0.44,
  fullCanAge = 10,
  MaxIntcptn = 0.17,
  LAImaxIntcptn = 3,
  alpha = 0.049,
  Y = 0.47,
  poolFractn = 0,
  e20 = 2.2,
  rhoAir = 1.2,
  lambda = 2460000,
  VPDconv = 0.000622,
  fracBB0 = 0.3,
  fracBB1 = 0.1,
  tBB = 10,
  rhoMin = 0.56,
  rhoMax = 0.56,
  tRho = 5,
  Qa = -50,
  Qb = 0.8,
  gDM_mol = 24,
  molPAR_MJ = 2.3,
  CoeffCond = 0.05,
  fCalpha700 = 1.433,
  fCg700 = 0.451,
  fCalphax = 2.33333333333333,
  fCg0 = 1.75,
  MinCond = 0,
  klmax = 0.06903041028657,
  krmax = 0.007021449543068,
  komax = 0.00840935020844,
  hc = 0.494539540825284,
  qir = 148.092408851735,
  qil = 14.9377246340992,
  qh = 40.4572399963215,
  qbc = 7.00578471546651,
  el = 0.739904866285297,
  er = 0.769592209202898,
  Nf = 0.0196791366844,
  Navm = 0.032129816636208,
  Navx = 8.75614468799907,
  leaf.grow = 4,
  leaf.fall = 11,
  Wl.s = 0.01,
  Wsbr.s = 0.1,
  Wr.s = 0.01,
  pWl.sprouts = 0.5,
  pWsbr.sprouts = 0.9,
  cod.pred = "3PG",
  cod.clim = "Month",
  waterBalanceSubMods = T,
  wiltPoint = 0.14,
  fieldCap = 0.2,
  satPoint = 0.35,
  K_s = 0.5,
  shared_area = 4,
  V_nr = 0.5,
  maxRootDepth = 0.5,
  sigma_zR = 0.3,
  SWC_nr = 10,
  E_S1 = 0.3,
  E_S2 = 0.6,
  MaxASW_state = 50,
  K_drain = 0.5,
  timeStp = 12,
  rainMod = 1,
  tempMod = 1,
  fsMod = 1,
  startN = 6,
  startC = 200,
  pseudo = F,
  Q10 = 4,
  Q10X = 0,
  Nleach_r = 0,
  K_drain_nrz = 0.8
)
}
\value{
list of parameters and values
}
\description{
getParmsOak get baseline parameters for oak (Species interactions under climate change in mixed stands of Scots pine and pedunculate oak Bouwman et al. 2021)
}
