% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_uncertainty.R
\name{NPPfunc}
\alias{NPPfunc}
\title{NPPfunc}
\usage{
NPPfunc(hzYrs, NPP_value, hazPeriod, grid_id, x, y, yc_value)
}
\arguments{
\item{hzYrs}{list of hazard years}

\item{NPP_value}{}

\item{hazPeriod}{}

\item{x}{x value}

\item{y}{y value}

\item{yc_value}{yield class}
}
\description{
takes hazard years, NPP_values and grid square data to return good and bad hazard years
}
