% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObsDevVals.R
\name{observedVals}
\alias{observedVals}
\title{observedVals get observed data and standard deviations for Harwood calibration}
\usage{
observedVals(timeStep, data, sY = 2015, eY = 2018, swc = F)
}
\arguments{
\item{timeStep}{time step to get observed data for}

\item{data}{Harwood data}

\item{sY}{start year for returned observed data}

\item{eY}{end year for returned observed data}

\item{swc}{whether to include soil water content in observed data}
}
\value{
vector of observed values
}
\description{
observedVals get observed data and standard deviations for Harwood calibration
}
