% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncsNew.R
\name{plotResultsNewMonthly}
\alias{plotResultsNewMonthly}
\title{plotResultsNewMonthly Function to plot the model against the Harwood data.}
\usage{
plotResultsNewMonthly(df, outSample, ShortTS = F, numSamps = 500)
}
\value{
pine plots
}
\description{
plotResultsNewMonthly Function to plot the model against the Harwood data.
}
