% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuncsNew.R
\name{plotResultsPine}
\alias{plotResultsPine}
\title{plotResultsPine Function to plot the model against the finnish data.}
\usage{
plotResultsPine(outP, ShortTS = F)
}
\value{
pine plots
}
\description{
plotResultsPine Function to plot the model against the finnish data.
}
