% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictVariablesInterest.r
\name{PredictVariablesInterest.3PG}
\alias{PredictVariablesInterest.3PG}
\title{PredictVariablesInterest.3PG predict key 3pg variables and update states}
\usage{
PredictVariablesInterest.3PG(state, parms, cod.pred, month)
}
\arguments{
\item{state}{current state of model}

\item{parms}{model parameters}

\item{cod.pred}{whether to use allometric predictions of dbh etc.}
}
\value{
updated model state and variables
}
\description{
PredictVariablesInterest.3PG predict key 3pg variables and update states
}
