% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPlot.R
\name{quickPlot}
\alias{quickPlot}
\title{quickPlot Function to plot the model against harwood.}
\usage{
quickPlot(flxdata_daily, output, grouping = "month")
}
\arguments{
\item{flxdata_daily}{harwood flux data}

\item{output}{model output}
}
\value{
pine plots
}
\description{
quickPlot Function to plot the model against harwood.
}
