% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_humidity_calc.R
\name{relative_humidity_calc}
\alias{relative_humidity_calc}
\title{relative_humidity_calc calc relative humidity}
\usage{
relative_humidity_calc(Tmean, Tref = 273.16, pp, spec_hum)
}
\arguments{
\item{Tmean}{mean temp K}

\item{Tref}{reference temperature (often 273.16)}

\item{spec_hum}{specific humidity}

\item{p}{pressure}
}
\value{
relative humidity
}
\description{
relative_humidity_calc calc relative humidity
}
