% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFuncsProjections.R
\name{soil_regr}
\alias{soil_regr}
\title{soil_regr}
\usage{
soil_regr(simDat, concChains)
}
\arguments{
\item{simDat}{current known parameters/variables}

\item{mcmcReg}{output from MCMC for fitting regional models}
}
\value{
simDat but with predicted soil texture params, es1,es2,swconst,swpower
}
\description{
use regression models to predict some unkown soil texture parameters from known parameters
}
