% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologyFuncs.R
\name{soilWC}
\alias{soilWC}
\title{soilWC soil water content calculator}
\usage{
soilWC(parms, weather, state, K_s, SWC, soilVol)
}
\arguments{
\item{parms}{global 3PG parameter values}

\item{weather}{global weather values}

\item{state}{state values}
}
\value{
SWC_rz SWC of rooting zone
}
\description{
Calculates soil water content for rooting zone
}
