% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoNWaL.r
\name{SoNWaL}
\alias{SoNWaL}
\title{SoNWaL function to run SoNWaL process based model}
\usage{
SoNWaL(
  weather,
  presc,
  t = 0,
  N = 2500,
  Wl = 0.01,
  WlDormant = 0,
  Wr = 0.01,
  Wsbr = 0.1,
  Wlitt = 0,
  YrC = 18.2,
  YlC = 6.5,
  OC = 122.77,
  YrN = 0,
  YlN = 1.52,
  ON = 6.9,
  Nav = 3,
  rotation = 1,
  cycle = 1,
  rm.sprouts = F,
  nyears = 35,
  initial.month = 1,
  latitude = 57.06,
  soilclass = -1,
  ASW = 200,
  MaxASW = 300,
  MinASW = 0,
  CO2 = 400,
  pFS2 = 0.76,
  pFS20 = 0.46,
  aS = 0.0572,
  nS = 2.4568,
  pRx = 0.48,
  pRn = 0.21,
  Tmin = -5,
  Topt = 15,
  Tmax = 35,
  kF = 1,
  SWconst0 = 0.7,
  SWpower0 = 9,
  m0 = 0.018,
  MaxAge = 265.6,
  nAge = 3.545,
  rAge = 0.796,
  gammaFx = 0.025,
  gammaF0 = 0.0022,
  tgammaF = 60,
  Rttover = 0.07,
  MaxCond = 0.02,
  LAIgcx = 3.33,
  BLcond = 0.2,
  wSx1000 = 255,
  thinPower = 1.5,
  mF = 0.2,
  mR = 0.2,
  mS = 0.2,
  SLA0 = 6,
  SLA1 = 4,
  tSLA = 6,
  k = 0.52,
  fullCanAge = 18,
  MaxIntcptn = 0.15,
  LAImaxIntcptn = 5,
  alpha = 0.05,
  Y = 0.49,
  poolFractn = 0,
  e20 = 2.2,
  rhoAir = 1.2,
  lambda = 2460000,
  VPDconv = 0.000622,
  fracBB0 = 0.3,
  fracBB1 = 0.1,
  tBB = 10,
  rhoMin = 0.39,
  rhoMax = 0.39,
  tRho = 5,
  Qa = -90,
  Qb = 0.8,
  gDM_mol = 24,
  molPAR_MJ = 2.3,
  CoeffCond = 0.05,
  fCalpha700 = 1.433,
  fCg700 = 0.451,
  fCalphax = 2.33333333333333,
  fCg0 = 1.75,
  MinCond = 0.015,
  klmax = 0.031532358,
  krmax = 0.0042394358,
  komax = 0.000458861,
  hc = 0.23,
  qir = 334.29,
  qil = 49.08,
  qh = 23.63,
  qbc = 2.21,
  el = 0.25,
  er = 0.56,
  Nf = 0.0684,
  Navm = 0.01,
  Navx = 20,
  leaf.grow = 0,
  leaf.fall = 0,
  Wl.s = 0.01,
  Wsbr.s = 0.1,
  Wr.s = 0.01,
  pWl.sprouts = 0.5,
  pWsbr.sprouts = 0.9,
  cod.pred = "3PG",
  cod.clim = "Average",
  K_s = 0.1,
  V_nr = 5,
  sigma_zR = 1,
  SWC_nr = 800,
  E_S = 100,
  E_S1 = 50,
  E_S2 = 0.01,
  waterBalanceSubMods = T,
  wiltPoint = 0.1,
  fieldCap = 0.3,
  MaxASW_state = 50,
  timeStp = 12,
  maxRootDepth = 2,
  shared_area = 5,
  K_drain = 0.1,
  SWC_rz,
  satPoint = 0.35,
  tempMod = 1,
  rainMod = 1,
  fsMod = 1,
  startN = 6,
  startC = 100,
  pseudo = F,
  Q10 = 0.5,
  Q10X = 2,
  Nleach_r = 0.5,
  K_drain_nrz
)
}
\value{
dataframe of SoNWaL output
}
\description{
SoNWaL function to run SoNWaL process based model
}
