% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoNWaL_spat_model_run.R
\name{SoNWaL_spat_model_run}
\alias{SoNWaL_spat_model_run}
\title{SoNWaL_spat_model_run execute run of SoNWaL for spatial grid cell with single parameter set}
\usage{
SoNWaL_spat_model_run(params, clm, scape = T)
}
\arguments{
\item{params}{vector of parameter set to run model with}

\item{clm}{climate data to run model with}
}
\value{
dataframe with model outputs, columns are output variables, rows are time-steps
}
\description{
SoNWaL_spat_model_run execute run of SoNWaL for spatial grid cell with single parameter set
}
