% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFuncsProjections.R
\name{spat_dat_scape}
\alias{spat_dat_scape}
\title{spat_dat_scape}
\usage{
spat_dat_scape(chunk = 1, output_dir, save_file)
}
\arguments{
\item{chunk}{chunk number, 1:39 covers whole of
scotland, can only go as high as files available from spatSplit}

\item{output_dir}{names of directory where files
(outputs from spatSplit) to go merge are located}

\item{save_file}{location to save merged files}
}
\value{
tibble of site id key with associated dataframe
of longitudinal climate data
}
\description{
read in list of CHESS-scape spatial
climate data with grid coordinates from spatSplit
function output and merge all climate vars into one
table associated with each grid cell
}
