% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologyFuncs.R
\name{SWGmod}
\alias{SWGmod}
\title{SWGmod soil water modifier (see landsberg and sands)}
\usage{
SWGmod(SWconst, SWpower, MoistRatio)
}
\value{
soil water modifier
}
\description{
SWGmod soil water modifier (see landsberg and sands)
}
