% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateASW.r
\name{UpdateASW}
\alias{UpdateASW}
\title{UpdateASW update hydrological state}
\usage{
UpdateASW(state, weather, site, parms, general.info)
}
\arguments{
\item{state}{current model state}

\item{weather}{weather data}

\item{site}{site data}

\item{parms}{model parameters}

\item{general.info}{some useful meta-data (legacy variable - can probably remove)}
}
\value{
updated model state
}
\description{
UpdateASW update hydrological state
}
