% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateSoil.r
\name{UpdateSoil}
\alias{UpdateSoil}
\title{UpdateSoil update soils state}
\usage{
UpdateSoil(state, parms, site, general.info, weather)
}
\arguments{
\item{state}{current model state}

\item{parms}{model parameters}

\item{site}{site data}

\item{general.info}{some useful meta-data (legacy variable - can probably remove)}

\item{weather}{weather data}
}
\value{
updated model state
}
\description{
UpdateSoil update soils state
}
