% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_class_calc.R
\name{yield_class_calc}
\alias{yield_class_calc}
\title{yield_class_calc: calculate yield class using MAI and CAI from SoNWaL output}
\usage{
yield_class_calc(out)
}
\arguments{
\item{out}{SoNWaL output}
}
\value{
yield class
}
\description{
yield_class_calc: calculate yield class using MAI and CAI from SoNWaL output
}
