% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialFuncsProjections.R
\name{add_BGS_dat}
\alias{add_BGS_dat}
\title{add_BGS_dat}
\usage{
add_BGS_dat(spat_chunk, soil_dat)
}
\arguments{
\item{spat_chunk}{spatial chunk to add soil data to}

\item{soil_dat}{location of BGS soil data files}
}
\value{
updated spat_chunk dataframe with additional soil data
}
\description{
add soil data from Astley (BGS data)
}
